unify fetching of a single implication
