+++ src/routes/category-implication/[id]/+page.svelte
@@ -1,10 +1,7 @@
 <script lang="ts">
 	import ImplicationPage from '$pages/ImplicationPage.svelte'
+
 	let { data } = $props()
 </script>
 
-<ImplicationPage
-	type="category"
-	implication={data.implication}
-	structures={data.categories}
-/>
+<ImplicationPage type="category" {...data} />
