+++ src/routes/functor-implication/[id]/+page.svelte
@@ -1,11 +1,7 @@
 <script lang="ts">
 	import ImplicationPage from '$pages/ImplicationPage.svelte'
+
 	let { data } = $props()
 </script>
 
-<ImplicationPage
-	type="functor"
-	associated_type="category"
-	implication={data.implication}
-	structures={data.functors}
-/>
+<ImplicationPage type="functor" associated_type="category" {...data} />
